added invincibility phase code
Now, when the player takes damage, they will have an invincibility phase that occurs for they duration that they are flashing. This allows enemies to have physical colliders of their own that the player can collide into and stop against.

On the player side, it now becomes much more obvious that hitting an enemy is bad.